day4_rag/q1_HR_RAG_MANUAL/backend/requirements.txt
flask
openai
faiss-cpu
python-dotenv
PyMuPDF
python-docx
numpy

